+++ src/nodely/engine/applicative/core_async.clj
@@ -2,36 +2,51 @@
   (:require
    [cats.protocols :as mp]
    [clojure.core.async :as async]
-   [clojure.core.async.impl.channels :as impl])
+   [clojure.core.async.impl.channels :as impl]
+   [nodely.data :as data]
+   [nodely.engine.applicative.protocols :as protocols]
+   [nodely.engine.core-async.core :as core-async.core])
   (:import
    [clojure.core.async.impl.channels ManyToManyChannel]))
 
 (declare context)
 
-(defn throw-err
+(defn nil-guard
+  [it]
+  (if (nil? it)
+    ::nil
+    it))
+
+(defn nil-unguard
+  [it]
+  (if (= it ::nil)
+    nil
+    it))
+
+(defn handle-read-value
   [v]
   (if (instance? Throwable v)
     (throw v)
-    v))
+    (nil-unguard v)))
 
 (defmacro <?
   "Just like a <! macro, but if the value taken from the channel is a Throwable instance, it will throw it.
   <! takes a val from port. Must be called inside a (go ...) or (go-try ...) block.
   Will return nil if closed. Will park if nothing is available."
   [ch]
-  `(throw-err (async/<! ~ch)))
+  `(handle-read-value (async/<! ~ch)))
 
 (extend-type ManyToManyChannel
   mp/Contextual
   (-get-context [_] context)
 
   mp/Extract
-  (-extract [it] (throw-err (async/<!! it))))
+  (-extract [it] (handle-read-value (async/<!! it))))
 
 (defn promise-of
   [v]
   (let [ret (async/promise-chan)]
-    (async/put! ret v)
+    (async/put! ret (nil-guard v))
     ret))
 
 (defmacro go-future
@@ -51,27 +66,37 @@
   exception will be immediately thrown."
   [& body]
   `(let [ret# (async/promise-chan)]
-     (async/go (let [retv# (try (do ~@body)
+     (async/go (let [retv# (try (nil-guard ~@body)
                                 (catch Throwable t#
                                   t#))]
                  (async/>! ret# retv#)))
      ret#))
 
-;; WIP Let's try abstracting the pattern we've been repeatedly doing
-;; BUT with error handling
-#_(defmacro go-promise
-    [& body]
-    (async/go (try ~@body)))
-
 (def ^:no-doc context
   (reify
     mp/Context
+    protocols/RunNode
+    (-apply-fn [_ f mv]
+      (let [tags (::data/tags (meta f))]
+        (cond (instance? nodely.engine.core_async.core.AsyncThunk f)
+              (go-future (let [in           (<? mv)
+                               exception-ch (async/promise-chan)
+                               result-ch    (core-async.core/evaluation-channel f in {:exception-ch exception-ch})
+                               merged-chs   (async/merge [result-ch exception-ch])
+                               result       (<? merged-chs)]
+                           (::data/value result)))
+              (contains? tags ::data/blocking)
+              (go-future (let [v (<? mv)]
+                           (async/thread (f v))))
+              :else
+              (go-future (let [v (<? mv)]
+                           (f v))))))
     mp/Functor
     (-fmap [mn f mv]
       (go-future (let [v (<? mv)]
                    (f v))))
 
-    mp/Monad ;; goood
+    mp/Monad
     (-mreturn [_ v]
       (promise-of v))